Adds routing to the angular app
This requires some changes to the way that we serve templates files to
work with the asset pipeline and moving the views into the
assets/javascripts directory and use of angular-rails-templates to
inject the templates into the JS via asset pipeline. This means that
all templates are now embedded in the JS rather than requiring
round-trips to the server to fetch them.

Client side tests are broken by this change because karma doesn't know
how to pre-process .html files to convert them to JS.